% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_indep_weights_object.R
\name{print.independence_weights}
\alias{print.independence_weights}
\title{Printing results for estimated energy balancing weights}
\usage{
\method{print}{independence_weights}(x, digits = max(getOption("digits") - 3, 3), ...)
}
\arguments{
\item{x}{a fitted object from \code{\link[independenceWeights]{independence_weights}}}

\item{digits}{minimal number of significant digits to print.}

\item{...}{further arguments passed to or from \code{\link[base]{print.default}}.}
}
\description{
Prints results for energy balancing weights
}
\seealso{
\code{\link[independenceWeights]{independence_weights}} for function which produces energy balancing weights
}
