% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/independence_weights_estimation.R
\name{weighted_energy_stats}
\alias{weighted_energy_stats}
\title{Calculation of weighted energy statistics for weighted dependence}
\usage{
weighted_energy_stats(A, X, weights, dimension_adj = FALSE, gamma = 1)
}
\arguments{
\item{A}{treatment vector indicating values of the treatment/exposure variable.}

\item{X}{matrix of covariates with number of rows equal to the length of \code{weights} and each column is a
covariate}

\item{weights}{a vector of sample weights}

\item{dimension_adj}{logical scalar. Whether or not to add adjustment to energy distance terms that account for
the dimensionality of \code{x}. Defaults to \code{TRUE}.}

\item{gamma}{positive numerical scalar. Defaults to 1 and should not be changed.}
}
\value{
a list with the following components
\item{ess}{The estimated effective sample size of the weights using Kish's effective sample size formula.}
\item{ess}{The estimated effective sample size of the weights using Kish's effective sample size formula.}
}
\description{
Calculates weighted energy statistics used to quantify weighted dependence
}
